RavenDB-17262 - Reducing large classes, starting to export methods to other classes in order to reduce nested class links in RVN cs.proj
