+++ src/Raven.Server/Commercial/SetupManager.cs
@@ -3,29 +3,20 @@ using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
 using System.Linq;
-using System.Net;
 using System.Net.Http;
-using System.Net.NetworkInformation;
 using System.Security.Cryptography.X509Certificates;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Connections;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json;
-using Newtonsoft.Json.Linq;
 using Raven.Client.Documents.Operations;
 using Raven.Client.ServerWide;
 using Raven.Client.ServerWide.Operations.Certificates;
 using Raven.Client.ServerWide.Operations.Configuration;
 using Raven.Client.Util;
+using Raven.Server.Commercial.LetsEncrypt;
 using Raven.Server.Config;
-using Raven.Server.Https;
 using Raven.Server.Json;
-using Raven.Server.Rachis;
 using Raven.Server.ServerWide;
 using Raven.Server.ServerWide.Commands;
 using Raven.Server.ServerWide.Context;
@@ -36,18 +27,14 @@ using Raven.Server.Web.Authentication;
 using Sparrow.Json;
 using Sparrow.Json.Parsing;
 using Sparrow.Logging;
-using Sparrow.Platform;
 using Sparrow.Server.Json.Sync;
 using Sparrow.Utils;
-using OpenFlags = System.Security.Cryptography.X509Certificates.OpenFlags;
 
 namespace Raven.Server.Commercial
 {
     public static class SetupManager
     {
-        private static readonly Logger Logger = LoggingSource.Instance.GetLogger<LicenseManager>("Server");
-        private const string GoogleDnsApi = "https://dns.google.com";
-
+        
         private static string BuildHostName(string nodeTag, string userDomain, string rootDomain)
         {
             return $"{nodeTag}.{userDomain}.{rootDomain}".ToLower();
@@ -76,11 +63,11 @@ namespace Raven.Server.Commercial
                 progress.AddInfo("Starting validation.");
                 onProgress(progress);
 
-                await ValidateSetupInfo(SetupMode.Secured, setupInfo, serverStore);
+                await LetsEncryptValidationApiHelper.ValidateSetupInfo(SetupMode.Secured, setupInfo, serverStore);
 
                 try
                 {
-                    await ValidateServerCanRunWithSuppliedSettings(setupInfo, serverStore, SetupMode.Secured, token);
+                    await LetsEncryptValidationApiHelper.ValidateServerCanRunWithSuppliedSettings(setupInfo, serverStore, SetupMode.Secured, token);
                 }
                 catch (Exception e)
                 {
@@ -147,8 +134,8 @@ namespace Raven.Server.Commercial
             if (Logger.IsOperationsEnabled)
                 Logger.Operations($"Successfully updated DNS record(s) and challenge(s) in {setupInfo.Domain.ToLower()}.{setupInfo.RootDomain.ToLower()}");
 
-            var cert = await LetsEncryptUtils.CompleteAuthorizationAndGetCertificate(
-                new LetsEncryptUtils.CompleteAuthorizationAndGetCertificateParameters
+            var cert = await LetsEncryptApiHelper.CompleteAuthorizationAndGetCertificate(
+                new LetsEncryptApiHelper.CompleteAuthorizationAndGetCertificateParameters
                 {
                     OnValidationSuccessful = () =>
                     {
@@ -220,7 +207,7 @@ namespace Raven.Server.Commercial
 
                     try
                     {
-                        await ValidateServerCanRunOnThisNode(settingsJsonObject, serverCert, serverStore, continueSetupInfo.NodeTag, token);
+                        await LetsEncryptValidationApiHelper.ValidateServerCanRunOnThisNode(settingsJsonObject, serverCert, serverStore, continueSetupInfo.NodeTag, token);
                     }
                     catch (Exception e)
                     {
@@ -386,127 +373,6 @@ namespace Raven.Server.Commercial
             return licenseStatus;
         }
 
-        public static async Task<IOperationResult> SetupLetsEncryptTask(Action<IOperationProgress> onProgress, SetupInfo setupInfo, ServerStore serverStore,
-            CancellationToken token)
-        {
-            var progress = new SetupProgressAndResult {Processed = 0, Total = 4};
-
-            try
-            {
-                var updatedLicense = new Reference<License>();
-                await GetUpdatedLicenseStatus(serverStore, setupInfo.License, updatedLicense).ConfigureAwait(false);
-                setupInfo.License = updatedLicense.Value;
-
-                AssertNoClusterDefined(serverStore);
-                progress.AddInfo("Setting up RavenDB in Let's Encrypt security mode.");
-                onProgress(progress);
-                try
-                {
-                    await ValidateSetupInfo(SetupMode.LetsEncrypt, setupInfo, serverStore);
-                }
-                catch (Exception e)
-                {
-                    throw new InvalidOperationException("Validation of supplied settings failed.", e);
-                }
-
-                progress.AddInfo($"Getting challenge(s) from Let's Encrypt. Using e-mail: {setupInfo.Email}.");
-                onProgress(progress);
-
-                var acmeClient = new LetsEncryptClient(serverStore.Configuration.Core.AcmeUrl);
-                await acmeClient.Init(setupInfo.Email, token);
-
-                var challengeResult = await LetsEncryptUtils.InitialLetsEncryptChallenge(setupInfo, acmeClient, token);
-
-                progress.Processed++;
-                progress.AddInfo(challengeResult.Challenge != null
-                    ? "Successfully received challenge(s) information from Let's Encrypt."
-                    : "Using cached Let's Encrypt certificate.");
-
-                progress.AddInfo($"Updating DNS record(s) and challenge(s) in {setupInfo.Domain.ToLower()}.{setupInfo.RootDomain.ToLower()}.");
-
-                onProgress(progress);
-
-                try
-                {
-                    await LetsEncryptUtils.UpdateDnsRecordsTask(new LetsEncryptUtils.UpdateDnsRecordParameters
-                    {
-                        OnProgress = onProgress,
-                        Progress = progress,
-                        Challenge = challengeResult.Challenge,
-                        SetupInfo = setupInfo,
-                        Token = token
-                    });
-                }
-                catch (Exception e)
-                {
-                    throw new InvalidOperationException(
-                        $"Failed to update DNS record(s) and challenge(s) in {setupInfo.Domain.ToLower()}.{setupInfo.RootDomain.ToLower()}", e);
-                }
-
-                progress.Processed++;
-                progress.AddInfo($"Successfully updated DNS record(s) and challenge(s) in {setupInfo.Domain.ToLower()}.{setupInfo.RootDomain.ToLower()}");
-                progress.AddInfo("Completing Let's Encrypt challenge(s)...");
-                onProgress(progress);
-
-                await LetsEncryptUtils.CompleteAuthorizationAndGetCertificate(new LetsEncryptUtils.CompleteAuthorizationAndGetCertificateParameters
-                {
-                    OnValidationSuccessful = () =>
-                    {
-                        progress.AddInfo("Let's Encrypt challenge(s) completed successfully.");
-                        progress.AddInfo("Acquiring certificate.");
-                        onProgress(progress);
-                    },
-                    SetupInfo = setupInfo,
-                    Client = acmeClient,
-                    ChallengeResult = challengeResult,
-                    ServerStore = serverStore,
-                    Token = token
-                });
-
-
-                progress.Processed++;
-                progress.AddInfo("Successfully acquired certificate from Let's Encrypt.");
-                progress.AddInfo("Starting validation.");
-                onProgress(progress);
-
-                try
-                {
-                    await ValidateServerCanRunWithSuppliedSettings(setupInfo, serverStore, SetupMode.LetsEncrypt, token);
-                }
-                catch (Exception e)
-                {
-                    throw new InvalidOperationException("Validation failed.", e);
-                }
-
-                progress.Processed++;
-                progress.AddInfo("Validation is successful.");
-                progress.AddInfo("Creating new RavenDB configuration settings.");
-
-                onProgress(progress);
-
-                try
-                {
-                    progress.SettingsZipFile =
-                        await CompleteClusterConfigurationAndGetSettingsZip(onProgress, progress, SetupMode.LetsEncrypt, setupInfo, serverStore, token);
-                }
-                catch (Exception e)
-                {
-                    throw new InvalidOperationException("Failed to create the configuration settings.", e);
-                }
-
-                progress.Processed++;
-                progress.AddInfo("Configuration settings created.");
-                progress.AddInfo("Setting up RavenDB in Let's Encrypt security mode finished successfully.");
-                onProgress(progress);
-            }
-            catch (Exception e)
-            {
-                LogErrorAndThrow(onProgress, progress, "Setting up RavenDB in Let's Encrypt security mode failed.", e);
-            }
-
-            return progress;
-        }
-
         private static void AssertNoClusterDefined(ServerStore serverStore)
         {
             var allNodes = serverStore.GetClusterTopology().AllNodes;
@@ -521,25 +387,6 @@ namespace Raven.Server.Commercial
             }
         }
 
-        private static async Task DeleteAllExistingCertificates(ServerStore serverStore)
-        {
-            // If a user repeats the setup process, there might be certificate leftovers in the cluster
-
-            List<string> existingCertificateKeys;
-            using (serverStore.ContextPool.AllocateOperationContext(out TransactionOperationContext context))
-            using (context.OpenReadTransaction())
-            {
-                existingCertificateKeys = serverStore.Cluster.GetCertificateThumbprintsFromCluster(context).ToList();
-            }
-
-            if (existingCertificateKeys.Count == 0)
-                return;
-
-            var res = await serverStore.SendToLeaderAsync(new DeleteCertificateCollectionFromClusterCommand(RaftIdGenerator.NewId()) {Names = existingCertificateKeys});
-
-            await serverStore.Cluster.WaitForIndexNotification(res.Index);
-        }
-
         private static void LogErrorAndThrow(Action<IOperationProgress> onProgress, SetupProgressAndResult progress, string msg, Exception e)
         {
             progress.AddError(msg, e);
@@ -547,332 +394,133 @@ namespace Raven.Server.Commercial
             throw new InvalidOperationException(msg, e);
         }
 
-        private static async Task UpdateDnsRecordsForCertificateRefreshTask(
-            string challenge,
-            SetupInfo setupInfo, CancellationToken token)
+
+
+
+    public static async Task<IOperationResult> SetupLetsEncryptTask(Action<IOperationProgress> onProgress, SetupInfo setupInfo, ServerStore serverStore,
+        CancellationToken token)
+    {
+        var progress = new SetupProgressAndResult
         {
-            using (var cts = CancellationTokenSource.CreateLinkedTokenSource(token, new CancellationTokenSource(TimeSpan.FromMinutes(15)).Token))
-            {
-                var registrationInfo = new RegistrationInfo
-                {
-                    License = setupInfo.License,
-                    Domain = setupInfo.Domain,
-                    Challenge = challenge,
-                    RootDomain = setupInfo.RootDomain,
-                    SubDomains = new List<RegistrationNodeInfo>()
-                };
+            Processed = 0,
+            Total = 4
+        };
 
-                foreach (var node in setupInfo.NodeSetupInfos)
-                {
-                    var regNodeInfo = new RegistrationNodeInfo {SubDomain = (node.Key + "." + setupInfo.Domain).ToLower(),};
-
-                    registrationInfo.SubDomains.Add(regNodeInfo);
-                }
-
-                var serializeObject = JsonConvert.SerializeObject(registrationInfo);
-
-                if (Logger.IsOperationsEnabled)
-                    Logger.Operations($"Start update process for certificate. License Id: {registrationInfo.License.Id}, " +
-                                      $"License Name: {registrationInfo.License.Name}, " +
-                                      $"Domain: {registrationInfo.Domain}, " +
-                                      $"RootDomain: {registrationInfo.RootDomain}");
-
-                HttpResponseMessage response;
-                try
-                {
-                    response = await ApiHttpClient.Instance.PostAsync("api/v1/dns-n-cert/register",
-                        new StringContent(serializeObject, Encoding.UTF8, "application/json"), token).ConfigureAwait(false);
-                }
-                catch (Exception e)
-                {
-                    throw new InvalidOperationException("Registration request to api.ravendb.net failed for: " + serializeObject, e);
-                }
-
-                var responseString = await response.Content.ReadAsStringAsync(cts.Token).ConfigureAwait(false);
-
-                if (response.IsSuccessStatusCode == false)
-                {
-                    throw new InvalidOperationException(
-                        $"Got unsuccessful response from registration request: {response.StatusCode}.{Environment.NewLine}{responseString}");
-                }
-
-                var id = (JsonConvert.DeserializeObject<Dictionary<string, string>>(responseString) ?? throw new InvalidOperationException()).First().Value;
-
-                try
-                {
-                    RegistrationResult registrationResult;
-                    do
-                    {
-                        try
-                        {
-                            await Task.Delay(1000, cts.Token);
-                            response = await ApiHttpClient.Instance.PostAsync("api/v1/dns-n-cert/registration-result?id=" + id,
-                                    new StringContent(serializeObject, Encoding.UTF8, "application/json"), cts.Token)
-                                .ConfigureAwait(false);
-                        }
-                        catch (Exception e)
-                        {
-                            throw new InvalidOperationException("Registration-result request to api.ravendb.net failed.", e); //add the object we tried to send to error
-                        }
-
-                        responseString = await response.Content.ReadAsStringAsync(cts.Token).ConfigureAwait(false);
-
-                        if (response.IsSuccessStatusCode == false)
-                        {
-                            throw new InvalidOperationException(
-                                $"Got unsuccessful response from registration-result request: {response.StatusCode}.{Environment.NewLine}{responseString}");
-                        }
-
-                        registrationResult = JsonConvert.DeserializeObject<RegistrationResult>(responseString);
-                    } while (registrationResult?.Status == "PENDING");
-                }
-                catch (Exception e)
-                {
-                    if (cts.IsCancellationRequested == false)
-                        throw;
-                    throw new TimeoutException("Request failed due to a timeout error", e);
-                }
-            }
-        }
-
-        private static async Task AssertLocalNodeCanListenToEndpoints(SetupInfo setupInfo, ServerStore serverStore)
+        try
         {
-            var localNode = setupInfo.NodeSetupInfos[setupInfo.LocalNodeTag];
-            var localIps = new List<IPEndPoint>();
+            var updatedLicense = new Reference<License>();
+            await LetsEncryptApiHelper.GetUpdatedLicenseStatus(serverStore, setupInfo.License, updatedLicense).ConfigureAwait(false);
+            setupInfo.License = updatedLicense.Value;
 
-            // Because we can get from user either an ip or a hostname, we resolve the hostname and get the actual ips it is mapped to
-            foreach (var hostnameOrIp in localNode.Addresses)
-            {
-                if (hostnameOrIp.Equals("0.0.0.0"))
-                {
-                    localIps.Add(new IPEndPoint(IPAddress.Parse(hostnameOrIp), localNode.Port));
-                    localIps.Add(new IPEndPoint(IPAddress.Parse(hostnameOrIp), localNode.TcpPort));
-                    continue;
-                }
-
-                foreach (var ip in await Dns.GetHostAddressesAsync(hostnameOrIp))
-                {
-                    localIps.Add(new IPEndPoint(IPAddress.Parse(ip.ToString()), localNode.Port));
-                    localIps.Add(new IPEndPoint(IPAddress.Parse(ip.ToString()), localNode.TcpPort));
-                }
-            }
-
-            var requestedEndpoints = localIps.ToArray();
-            var currentServerEndpoints = serverStore.Server.ListenEndpoints.Addresses.Select(ip => new IPEndPoint(ip, serverStore.Server.ListenEndpoints.Port)).ToList();
-
-            var ipProperties = IPGlobalProperties.GetIPGlobalProperties();
-            IPEndPoint[] activeTcpListeners;
+            AssertNoClusterDefined(serverStore);
+            progress.AddInfo("Setting up RavenDB in Let's Encrypt security mode.");
+            onProgress(progress);
             try
             {
-                activeTcpListeners = ipProperties.GetActiveTcpListeners();
-            }
-            catch (Exception)
-            {
-                // If GetActiveTcpListeners is not supported, skip the validation
-                // See https://github.com/dotnet/corefx/issues/30909
-                return;
-            }
-
-            foreach (var requestedEndpoint in requestedEndpoints)
-            {
-                if (activeTcpListeners.Contains(requestedEndpoint))
-                {
-                    if (currentServerEndpoints.Contains(requestedEndpoint))
-                        continue; // OK... used by the current server
-
-                    throw new InvalidOperationException(
-                        $"The requested endpoint '{requestedEndpoint.Address}:{requestedEndpoint.Port}' is already in use by another process. You may go back in the wizard, change the settings and try again.");
-                }
-            }
-        }
-
-        private static async Task ValidateServerCanRunWithSuppliedSettings(SetupInfo setupInfo, ServerStore serverStore, SetupMode setupMode, CancellationToken token)
-        {
-            var localNode = setupInfo.NodeSetupInfos[setupInfo.LocalNodeTag];
-            var localIps = new List<IPEndPoint>();
-
-            foreach (var hostnameOrIp in localNode.Addresses)
-            {
-                if (hostnameOrIp.Equals("0.0.0.0"))
-                {
-                    localIps.Add(new IPEndPoint(IPAddress.Parse(hostnameOrIp), localNode.Port));
-                    continue;
-                }
-
-                foreach (var ip in await Dns.GetHostAddressesAsync(hostnameOrIp))
-                {
-                    localIps.Add(new IPEndPoint(IPAddress.Parse(ip.ToString()), localNode.Port));
-                }
-            }
-
-            var serverCert = setupInfo.GetX509Certificate();
-
-            var localServerUrl =
-                LetsEncryptUtils.GetServerUrlFromCertificate(serverCert, setupInfo, setupInfo.LocalNodeTag, localNode.Port, localNode.TcpPort, out _, out _);
-
-            try
-            {
-                if (serverStore.Server.ListenEndpoints.Port == localNode.Port)
-                {
-                    var currentIps = serverStore.Server.ListenEndpoints.Addresses.ToList();
-
-                    if (localIps.Count == 0 && currentIps.Count == 1 &&
-                        (Equals(currentIps[0], IPAddress.Any) || Equals(currentIps[0], IPAddress.IPv6Any)))
-                        return; // listen to any ip in this
-
-                    if (localIps.All(ip => currentIps.Contains(ip.Address)))
-                        return; // we already listen to all these IPs, no need to check
-                }
-
-                if (setupMode == SetupMode.LetsEncrypt)
-                {
-                    // In case an external ip was specified, this is the ip we update in the dns records. (not the one we bind to)
-                    var ips = localNode.ExternalIpAddress == null
-                        ? localIps.ToArray()
-                        : new[] {new IPEndPoint(IPAddress.Parse(localNode.ExternalIpAddress), localNode.ExternalPort)};
-
-                    await AssertDnsUpdatedSuccessfully(localServerUrl, ips, token);
-                }
-
-                // Here we send the actual ips we will bind to in the local machine.
-                await SimulateRunningServer(serverStore, serverCert, localServerUrl, setupInfo.LocalNodeTag, localIps.ToArray(), localNode.Port,
-                    serverStore.Configuration.ConfigPath, setupMode, token);
+                await ValidateSetupInfo(SetupMode.LetsEncrypt, setupInfo, serverStore);
             }
             catch (Exception e)
             {
-                throw new InvalidOperationException("Failed to simulate running the server with the supplied settings using: " + localServerUrl, e);
+                throw new InvalidOperationException("Validation of supplied settings failed.", e);
             }
-        }
 
-        private static async Task ValidateServerCanRunOnThisNode(BlittableJsonReaderObject settingsJsonObject, X509Certificate2 cert, ServerStore serverStore,
-            string nodeTag, CancellationToken token)
-        {
-            settingsJsonObject.TryGet(RavenConfiguration.GetKey(x => x.Core.PublicServerUrl), out string publicServerUrl);
-            settingsJsonObject.TryGet(RavenConfiguration.GetKey(x => x.Core.ServerUrls), out string serverUrl);
-            settingsJsonObject.TryGet(RavenConfiguration.GetKey(x => x.Core.SetupMode), out SetupMode setupMode);
-            settingsJsonObject.TryGet(RavenConfiguration.GetKey(x => x.Core.ExternalIp), out string externalIp);
+            progress.AddInfo($"Getting challenge(s) from Let's Encrypt. Using e-mail: {setupInfo.Email}.");
+            onProgress(progress);
 
-            var serverUrls = serverUrl.Split(";");
-            var port = new Uri(serverUrls[0]).Port;
-            var hostnamesOrIps = serverUrls.Select(url => new Uri(url).DnsSafeHost).ToArray();
+            var acmeClient = new LetsEncryptClient(serverStore.Configuration.Core.AcmeUrl);
+            await acmeClient.Init(setupInfo.Email, token);
 
-            var localIps = new List<IPEndPoint>();
+            var challengeResult = await LetsEncryptUtils.InitialLetsEncryptChallenge(setupInfo, acmeClient, token);
 
-            foreach (var hostnameOrIp in hostnamesOrIps)
-            {
-                if (hostnameOrIp.Equals("0.0.0.0"))
-                {
-                    localIps.Add(new IPEndPoint(IPAddress.Parse(hostnameOrIp), port));
-                    continue;
-                }
+            progress.Processed++;
+            progress.AddInfo(challengeResult.Challenge != null
+                ? "Successfully received challenge(s) information from Let's Encrypt."
+                : "Using cached Let's Encrypt certificate.");
 
-                foreach (var ip in await Dns.GetHostAddressesAsync(hostnameOrIp))
-                {
-                    localIps.Add(new IPEndPoint(IPAddress.Parse(ip.ToString()), port));
-                }
-            }
+            progress.AddInfo($"Updating DNS record(s) and challenge(s) in {setupInfo.Domain.ToLower()}.{setupInfo.RootDomain.ToLower()}.");
+
+            onProgress(progress);
 
             try
             {
-                if (serverStore.Server.ListenEndpoints.Port == port)
+                await RavenDnsRecordHelper.UpdateDnsRecordsTask(new RavenDnsRecordHelper.UpdateDnsRecordParameters
                 {
-                    var currentIps = serverStore.Server.ListenEndpoints.Addresses.ToList();
-
-                    if (localIps.Count == 0 && currentIps.Count == 1 &&
-                        (Equals(currentIps[0], IPAddress.Any) || Equals(currentIps[0], IPAddress.IPv6Any)))
-                        return; // listen to any ip in this
-
-                    if (localIps.All(ip => currentIps.Contains(ip.Address)))
-                        return; // we already listen to all these IPs, no need to check
-                }
-
-                if (setupMode == SetupMode.LetsEncrypt)
-                {
-                    var ips = string.IsNullOrEmpty(externalIp)
-                        ? localIps.ToArray()
-                        : new[] {new IPEndPoint(IPAddress.Parse(externalIp), port)};
-
-                    await AssertDnsUpdatedSuccessfully(publicServerUrl, ips, token);
-                }
-
-                // Here we send the actual ips we will bind to in the local machine.
-                await SimulateRunningServer(serverStore, cert, publicServerUrl, nodeTag, localIps.ToArray(), port, serverStore.Configuration.ConfigPath, setupMode,
-                    token);
+                    OnProgress = onProgress,
+                    Progress = progress,
+                    Challenge = challengeResult.Challenge,
+                    SetupInfo = setupInfo,
+                    Token = token
+                });
             }
             catch (Exception e)
             {
-                throw new InvalidOperationException("Failed to simulate running the server with the supplied settings using: " + publicServerUrl, e);
+                throw new InvalidOperationException(
+                    $"Failed to update DNS record(s) and challenge(s) in {setupInfo.Domain.ToLower()}.{setupInfo.RootDomain.ToLower()}", e);
             }
-        }
 
-        private static async Task ValidateSetupInfo(SetupMode setupMode, SetupInfo setupInfo, ServerStore serverStore)
-        {
-            if ((await SetupParameters.Get(serverStore)).IsDocker)
+            progress.Processed++;
+            progress.AddInfo($"Successfully updated DNS record(s) and challenge(s) in {setupInfo.Domain.ToLower()}.{setupInfo.RootDomain.ToLower()}");
+            progress.AddInfo("Completing Let's Encrypt challenge(s)...");
+            onProgress(progress);
+
+            await LetsEncryptApiHelper.CompleteAuthorizationAndGetCertificate(new LetsEncryptUtils.CompleteAuthorizationAndGetCertificateParameters
             {
-                if (setupInfo.NodeSetupInfos[setupInfo.LocalNodeTag].Addresses.Any(ip => ip.StartsWith("127.")))
+                OnValidationSuccessful = () =>
                 {
-                    throw new InvalidOperationException("When the server is running in Docker, you cannot bind to ip 127.X.X.X, please use the hostname instead.");
-                }
-            }
+                    progress.AddInfo("Let's Encrypt challenge(s) completed successfully.");
+                    progress.AddInfo("Acquiring certificate.");
+                    onProgress(progress);
+                },
+                SetupInfo = setupInfo,
+                Client = acmeClient,
+                ChallengeResult = challengeResult,
+                ServerStore = serverStore,
+                Token = token
+            });
 
-            switch (setupMode)
-            {
-                case SetupMode.LetsEncrypt when setupInfo.NodeSetupInfos.ContainsKey(setupInfo.LocalNodeTag) == false:
-                    throw new ArgumentException($"At least one of the nodes must have the node tag '{setupInfo.LocalNodeTag}'.");
-                case SetupMode.LetsEncrypt when IsValidEmail(setupInfo.Email) == false:
-                    throw new ArgumentException("Invalid email address.");
-                case SetupMode.LetsEncrypt when IsValidDomain(setupInfo.Domain + "." + setupInfo.RootDomain) == false:
-                    throw new ArgumentException("Invalid domain name.");
-                case SetupMode.LetsEncrypt when setupInfo.ClientCertNotAfter.HasValue && setupInfo.ClientCertNotAfter <= DateTime.UtcNow.Date:
-                    throw new ArgumentException("The client certificate expiration date must be in the future.");
-                case SetupMode.Secured when string.IsNullOrWhiteSpace(setupInfo.Certificate):
-                    throw new ArgumentException($"{nameof(setupInfo.Certificate)} is a mandatory property for a secured setup");
-            }
 
-            foreach ((string key, SetupInfo.NodeInfo value) in setupInfo.NodeSetupInfos)
-            {
-                RachisConsensus.ValidateNodeTag(key);
+            progress.Processed++;
+            progress.AddInfo("Successfully acquired certificate from Let's Encrypt.");
+            progress.AddInfo("Starting validation.");
+            onProgress(progress);
 
-                if (value.Port == 0)
-                    setupInfo.NodeSetupInfos[key].Port = 443;
-
-                if (value.TcpPort == 0)
-                    setupInfo.NodeSetupInfos[key].TcpPort = 38888;
-
-                if (setupMode == SetupMode.LetsEncrypt &&
-                    setupInfo.NodeSetupInfos[key].Addresses.Any(ip => ip.Equals("0.0.0.0")) &&
-                    string.IsNullOrWhiteSpace(setupInfo.NodeSetupInfos[key].ExternalIpAddress))
-                {
-                    throw new ArgumentException("When choosing 0.0.0.0 as the ip address, you must provide an external ip to update in the DNS records.");
-                }
-            }
-
-            await AssertLocalNodeCanListenToEndpoints(setupInfo, serverStore);
-        }
-
-        public static bool IsValidEmail(string email)
-        {
-            if (string.IsNullOrWhiteSpace(email))
-                return false;
             try
             {
-                var address = new System.Net.Mail.MailAddress(email);
-                return address.Address == email;
+                await ValidateServerCanRunWithSuppliedSettings(setupInfo, serverStore, SetupMode.LetsEncrypt, token);
             }
-            catch
+            catch (Exception e)
             {
-                return false;
+                throw new InvalidOperationException("Validation failed.", e);
             }
-        }
 
-        private static bool IsValidDomain(string domain)
+            progress.Processed++;
+            progress.AddInfo("Validation is successful.");
+            progress.AddInfo("Creating new RavenDB configuration settings.");
+
+            onProgress(progress);
+
+            try
+            {
+                progress.SettingsZipFile =
+                    await CompleteClusterConfigurationAndGetSettingsZip(onProgress, progress, SetupMode.LetsEncrypt, setupInfo, serverStore, token);
+            }
+            catch (Exception e)
+            {
+                throw new InvalidOperationException("Failed to create the configuration settings.", e);
+            }
+
+            progress.Processed++;
+            progress.AddInfo("Configuration settings created.");
+            progress.AddInfo("Setting up RavenDB in Let's Encrypt security mode finished successfully.");
+            onProgress(progress);
+        }
+        catch (Exception e)
         {
-            if (string.IsNullOrWhiteSpace(domain))
-                return false;
-
-            return Uri.CheckHostName(domain) != UriHostNameType.Unknown;
+            LogErrorAndThrow(onProgress, progress, "Setting up RavenDB in Let's Encrypt security mode failed.", e);
         }
 
+        return progress;
+    }
 
         private static async Task CompleteConfigurationForNewNode(
             Action<IOperationProgress> onProgress,
@@ -911,7 +559,7 @@ namespace Raven.Server.Commercial
             {
                 await serverStore.EnsureNotPassiveAsync(publicServerUrl, firstNodeTag);
 
-                await DeleteAllExistingCertificates(serverStore);
+                await LetsEncryptCertificateUtil.DeleteAllExistingCertificates(serverStore);
 
                 if (setupMode == SetupMode.LetsEncrypt && license != null)
                 {
@@ -1035,8 +683,8 @@ namespace Raven.Server.Commercial
                 progress.AddInfo($"Saving configuration at {serverStore.Configuration.ConfigPath}.");
                 onProgress(progress);
 
-                var indentedJson = LetsEncryptUtils.IndentJsonString(settingsJsonObject.ToString());
-                LetsEncryptUtils.WriteSettingsJsonLocally(serverStore.Configuration.ConfigPath, indentedJson);
+                var indentedJson = LetsEncryptApiHelper.IndentJsonString(settingsJsonObject.ToString());
+                LetsEncryptApiHelper.WriteSettingsJsonLocally(serverStore.Configuration.ConfigPath, indentedJson);
             }
             catch (Exception e)
             {
@@ -1045,15 +693,14 @@ namespace Raven.Server.Commercial
 
             try
             {
-                progress.Readme = LetsEncryptUtils.CreateReadmeText(continueSetupInfo.NodeTag, publicServerUrl, true, continueSetupInfo.RegisterClientCert);
+                progress.Readme = LetsEncryptApiHelper.CreateReadmeText(continueSetupInfo.NodeTag, publicServerUrl, true, continueSetupInfo.RegisterClientCert);
             }
             catch (Exception e)
             {
                 throw new InvalidOperationException("Failed to create the readme text.", e);
             }
         }
-
-        private static async Task<byte[]> CompleteClusterConfigurationAndGetSettingsZip(
+     private static async Task<byte[]> CompleteClusterConfigurationAndGetSettingsZip(
             Action<IOperationProgress> onProgress,
             SetupProgressAndResult progress,
             SetupMode setupMode,
@@ -1125,6 +772,78 @@ namespace Raven.Server.Commercial
                 },
             });
         }
+        // private static async Task<byte[]> CompleteClusterConfigurationAndGetSettingsZip(
+        //     Action<IOperationProgress> onProgress,
+        //     SetupProgressAndResult progress,
+        //     SetupMode setupMode,
+        //     SetupInfo setupInfo,
+        //     ServerStore serverStore,
+        //     CancellationToken token)
+        // {
+        //     return await LetsEncryptUtils.CompleteClusterConfigurationAndGetSettingsZip(new LetsEncryptUtils.CompleteClusterConfigurationParameters
+        //     {
+        //         OnProgress = onProgress,
+        //         Progress = progress,
+        //         SetupInfo = setupInfo,
+        //         SettingsPath = serverStore.Configuration.ConfigPath,
+        //         OnWriteSettingsJsonLocally = indentedJson =>
+        //         {
+        //             return Task.Run(() => LetsEncryptUtils.WriteSettingsJsonLocally(serverStore.Configuration.ConfigPath, indentedJson), token);
+        //         },
+        //         OnGetCertificatePath = certificateFileName =>
+        //         {
+        //             return Task.Run(() => serverStore.Configuration.GetSetting(
+        //                                       RavenConfiguration.GetKey(x => x.Core.SetupResultingServerCertificatePath))
+        //                                   ?? Path.Combine(AppContext.BaseDirectory, certificateFileName), token);
+        //         },
+        //         OnPutServerWideStudioConfigurationValues = async studioEnvironment =>
+        //         {
+        //             var res = await serverStore.PutValueInClusterAsync(new PutServerWideStudioConfigurationCommand(
+        //                 new ServerWideStudioConfiguration {Disabled = false, Environment = studioEnvironment}, RaftIdGenerator.DontCareId));
+        //
+        //             await serverStore.Cluster.WaitForIndexNotification(res.Index);
+        //         },
+        //         OnBeforeAddingNodesToCluster = async (publicServerUrl, localNodeTag) =>
+        //         {
+        //             try
+        //             {
+        //                 serverStore.Engine.SetNewState(RachisState.Passive, null, serverStore.Engine.CurrentTerm, "During setup wizard, " +
+        //                     "making sure there is no cluster from previous installation.");
+        //             }
+        //             catch (Exception e)
+        //             {
+        //                 throw new InvalidOperationException("Failed to delete previous cluster topology during setup.", e);
+        //             }
+        //
+        //             await serverStore.EnsureNotPassiveAsync(publicServerUrl, setupInfo.LocalNodeTag);
+        //
+        //             await DeleteAllExistingCertificates(serverStore);
+        //
+        //             if (setupMode == SetupMode.LetsEncrypt)
+        //             {
+        //                 await serverStore.EnsureNotPassiveAsync(skipLicenseActivation: true);
+        //                 await serverStore.LicenseManager.ActivateAsync(setupInfo.License, RaftIdGenerator.DontCareId);
+        //             }
+        //
+        //             serverStore.HasFixedPort = setupInfo.NodeSetupInfos[localNodeTag].Port != 0;
+        //         },
+        //         AddNodeToCluster = async nodeTag =>
+        //         {
+        //             try
+        //             {
+        //                 await serverStore.AddNodeToClusterAsync(setupInfo.NodeSetupInfos[nodeTag].PublicServerUrl, nodeTag, validateNotInTopology: false, token: token);
+        //             }
+        //             catch (Exception e)
+        //             {
+        //                 throw new InvalidOperationException($"Failed to add node '{nodeTag}' to the cluster.", e);
+        //             }
+        //         },
+        //         RegisterClientCertInOs = (onProgressCopy, progressCopy, clientCert) =>
+        //         {
+        //             return Task.Run(() => RegisterClientCertInOs(onProgressCopy, progressCopy, clientCert), token);
+        //         },
+        //     });
+        // }
 
 
         private static void RegisterClientCertInOs(Action<IOperationProgress> onProgress, SetupProgressAndResult progress, X509Certificate2 clientCert)
@@ -1145,241 +864,8 @@ namespace Raven.Server.Commercial
             }
         }
 
+      
 
-        private class UniqueResponseResponder : IStartup
-        {
-            private readonly string _response;
-
-            public UniqueResponseResponder(string response)
-            {
-                _response = response;
-            }
-
-            public IServiceProvider ConfigureServices(IServiceCollection services)
-            {
-                return services.BuildServiceProvider();
-            }
-
-            public void Configure(IApplicationBuilder app)
-            {
-                app.Run(async context =>
-                {
-                    await context.Response.WriteAsync(_response);
-                });
-            }
-        }
-
-        private static async Task SimulateRunningServer(ServerStore serverStore, X509Certificate2 serverCertificate, string serverUrl, string nodeTag,
-            IPEndPoint[] addresses, int port, string settingsPath, SetupMode setupMode, CancellationToken token)
-        {
-            var configuration = RavenConfiguration.CreateForServer(null, settingsPath);
-            configuration.Initialize();
-            var guid = Guid.NewGuid().ToString();
-
-            IWebHost webHost = null;
-            try
-            {
-                try
-                {
-                    var responder = new UniqueResponseResponder(guid);
-
-                    var webHostBuilder = new WebHostBuilder()
-                        .CaptureStartupErrors(captureStartupErrors: true)
-                        .UseKestrel(options =>
-                        {
-                            var httpsConnectionMiddleware = new HttpsConnectionMiddleware(serverStore.Server, options);
-                            httpsConnectionMiddleware.SetCertificate(serverCertificate);
-
-                            if (addresses.Length == 0)
-                            {
-                                var defaultIp = new IPEndPoint(IPAddress.Parse("0.0.0.0"), port == 0 ? 443 : port);
-
-                                options.Listen(defaultIp, listenOptions =>
-                                {
-                                    listenOptions
-                                        .UseHttps()
-                                        .Use(httpsConnectionMiddleware.OnConnectionAsync);
-                                });
-                                if (Logger.IsInfoEnabled)
-                                    Logger.Info($"List of ip addresses for node '{nodeTag}' is empty. WebHost listening to {defaultIp}");
-                            }
-
-                            foreach (var address in addresses)
-                            {
-                                options.Listen(address, listenOptions =>
-                                {
-                                    listenOptions
-                                        .UseHttps()
-                                        .Use(httpsConnectionMiddleware.OnConnectionAsync);
-                                });
-                            }
-                        })
-                        .UseSetting(WebHostDefaults.ApplicationKey, "Setup simulation")
-                        .ConfigureServices(collection => { collection.AddSingleton(typeof(IStartup), responder); })
-                        .UseShutdownTimeout(TimeSpan.FromMilliseconds(150));
-
-                    webHost = webHostBuilder.Build();
-
-                    await webHost.StartAsync(token);
-                }
-                catch (Exception e)
-                {
-                    string linuxMsg = null;
-                    if (PlatformDetails.RunningOnPosix && (port == 80 || port == 443))
-                    {
-                        linuxMsg = $"It can happen if port '{port}' is not allowed for the non-root RavenDB process." +
-                                   $"Try using setcap to allow it: sudo setcap CAP_NET_BIND_SERVICE=+eip {Path.Combine(AppContext.BaseDirectory, "Raven.Server")}";
-                    }
-
-                    var also = linuxMsg == null ? string.Empty : "also";
-                    var externalIpMsg = setupMode == SetupMode.LetsEncrypt
-                        ? $"It can {also} happen if the ip is external (behind a firewall, docker). If this is the case, try going back to the previous screen and add the same ip as an external ip."
-                        : string.Empty;
-
-                    throw new InvalidOperationException(
-                        $"Failed to start WebHost on node '{nodeTag}'. The specified ip address might not be reachable due to network issues. {linuxMsg}{Environment.NewLine}{externalIpMsg}{Environment.NewLine}" +
-                        $"Settings file:{settingsPath}.{Environment.NewLine}" +
-                        $"IP addresses: {string.Join(", ", addresses.Select(address => address.ToString()))}.", e);
-                }
-
-                using (var httpMessageHandler = new HttpClientHandler())
-                {
-                    // on MacOS this is not supported because Apple...
-                    if (PlatformDetails.RunningOnMacOsx == false)
-                    {
-                        httpMessageHandler.ServerCertificateCustomValidationCallback += (_, certificate2, _, _) =>
-                            // we want to verify that we get the same thing back
-                        {
-                            if (certificate2.Thumbprint != serverCertificate.Thumbprint)
-                                throw new InvalidOperationException("Expected to get " + serverCertificate.FriendlyName + " with thumbprint " +
-                                                                    serverCertificate.Thumbprint + " but got " +
-                                                                    certificate2.FriendlyName + " with thumbprint " + certificate2.Thumbprint);
-                            return true;
-                        };
-                    }
-
-                    using (var client = new HttpClient(httpMessageHandler) {BaseAddress = new Uri(serverUrl),})
-                    {
-                        HttpResponseMessage response = null;
-                        string result = null;
-                        try
-                        {
-                            using (var cancellationTokenSource = new CancellationTokenSource(TimeSpan.FromSeconds(30)))
-                            using (var cts = CancellationTokenSource.CreateLinkedTokenSource(token, cancellationTokenSource.Token))
-                            {
-                                response = await client.GetAsync("/are-you-there?", cts.Token);
-                                response.EnsureSuccessStatusCode();
-                                result = await response.Content.ReadAsStringAsync(cts.Token);
-                                if (result != guid)
-                                {
-                                    throw new InvalidOperationException($"Expected result guid: {guid} but got {result}.");
-                                }
-                            }
-                        }
-                        catch (Exception e)
-                        {
-                            if (setupMode == SetupMode.Secured && await CanResolveHostNameLocally(serverUrl, addresses) == false)
-                            {
-                                throw new InvalidOperationException(
-                                    $"Failed to resolve '{serverUrl}'. Try to clear your local/network DNS cache and restart validation.", e);
-                            }
-
-                            throw new InvalidOperationException($"Client failed to contact WebHost listening to '{serverUrl}'.{Environment.NewLine}" +
-                                                                $"Are you blocked by a firewall? Make sure the port is open.{Environment.NewLine}" +
-                                                                $"Settings file:{settingsPath}.{Environment.NewLine}" +
-                                                                $"IP addresses: {string.Join(", ", addresses.Select(address => address.ToString()))}.{Environment.NewLine}" +
-                                                                $"Response: {response?.StatusCode}.{Environment.NewLine}{result}", e);
-                        }
-                    }
-                }
-            }
-            finally
-            {
-                if (webHost != null)
-                    await webHost.StopAsync(TimeSpan.Zero);
-            }
-        }
-
-        private static async Task<bool> CanResolveHostNameLocally(string serverUrl, IPEndPoint[] expectedAddresses)
-        {
-            var expectedIps = expectedAddresses.Select(address => address.Address.ToString()).ToHashSet();
-            var hostname = new Uri(serverUrl).Host;
-            HashSet<string> actualIps;
-
-            try
-            {
-                actualIps = (await Dns.GetHostAddressesAsync(hostname)).Select(address => address.ToString()).ToHashSet();
-            }
-            catch (Exception)
-            {
-                return false;
-            }
-
-            return expectedIps.SetEquals(actualIps);
-        }
-
-        private static async Task AssertDnsUpdatedSuccessfully(string serverUrl, IPEndPoint[] expectedAddresses, CancellationToken token)
-        {
-            // First we'll try to resolve the hostname through google's public dns api
-            using (var cancellationTokenSource = new CancellationTokenSource(TimeSpan.FromSeconds(30)))
-            using (var cts = CancellationTokenSource.CreateLinkedTokenSource(token, cancellationTokenSource.Token))
-            {
-                var expectedIps = expectedAddresses.Select(address => address.Address.ToString()).ToHashSet();
-
-                var hostname = new Uri(serverUrl).Host;
-
-                using (var client = new HttpClient {BaseAddress = new Uri(GoogleDnsApi)})
-                {
-                    var response = await client.GetAsync($"/resolve?name={hostname}", cts.Token);
-
-                    var responseString = await response.Content.ReadAsStringAsync(cts.Token).ConfigureAwait(false);
-                    if (response.IsSuccessStatusCode == false)
-                        throw new InvalidOperationException($"Tried to resolve '{hostname}' using Google's api ({GoogleDnsApi}).{Environment.NewLine}"
-                                                            + $"Request failed with status {response.StatusCode}.{Environment.NewLine}{responseString}");
-
-                    dynamic dnsResult = JsonConvert.DeserializeObject(responseString);
-
-                    // DNS response format: https://developers.google.com/speed/public-dns/docs/dns-over-https
-
-                    if (dnsResult?.Status != 0)
-                        throw new InvalidOperationException($"Tried to resolve '{hostname}' using Google's api ({GoogleDnsApi}).{Environment.NewLine}"
-                                                            + $"Got a DNS failure response:{Environment.NewLine}{responseString}" +
-                                                            Environment.NewLine +
-                                                            "Please wait a while until DNS propagation is finished and try again. If you are trying to update existing DNS records, it might take hours to update because of DNS caching. If the issue persists, contact RavenDB's support.");
-
-                    JArray answers = dnsResult.Answer;
-                    var googleIps = answers.Select(answer => answer["data"].ToString()).ToHashSet();
-
-                    if (googleIps.SetEquals(expectedIps) == false)
-                        throw new InvalidOperationException($"Tried to resolve '{hostname}' using Google's api ({GoogleDnsApi}).{Environment.NewLine}"
-                                                            + $"Expected to get these ips: {string.Join(", ", expectedIps)} while Google's actual result was: {string.Join(", ", googleIps)}"
-                                                            + Environment.NewLine +
-                                                            "Please wait a while until DNS propagation is finished and try again. If you are trying to update existing DNS records, it might take hours to update because of DNS caching. If the issue persists, contact RavenDB's support.");
-                }
-
-                // Resolving through google worked, now let's check locally
-                HashSet<string> actualIps;
-                try
-                {
-                    actualIps = (await Dns.GetHostAddressesAsync(hostname)).Select(address => address.ToString()).ToHashSet();
-                }
-                catch (Exception e)
-                {
-                    throw new InvalidOperationException(
-                        $"Cannot resolve '{hostname}' locally but succeeded resolving the address using Google's api ({GoogleDnsApi})."
-                        + Environment.NewLine + "Try to clear your local/network DNS cache or wait a few minutes and try again."
-                        + Environment.NewLine + "Another temporary solution is to configure your local network connection to use Google's DNS server (8.8.8.8).", e);
-                }
-
-                if (expectedIps.SetEquals(actualIps) == false)
-                    throw new InvalidOperationException(
-                        $"Tried to resolve '{hostname}' locally but got an outdated result."
-                        + Environment.NewLine + $"Expected to get these ips: {string.Join(", ", expectedIps)} while the actual result was: {string.Join(", ", actualIps)}"
-                        + Environment.NewLine + $"If we try resolving through Google's api ({GoogleDnsApi}), it works well."
-                        + Environment.NewLine + "Try to clear your local/network DNS cache or wait a few minutes and try again."
-                        + Environment.NewLine + "Another temporary solution is to configure your local network connection to use Google's DNS server (8.8.8.8).");
-            }
-        }
         
         public static async Task<byte[]> GenerateCertificateTask(string name, ServerStore serverStore, SetupInfo setupInfo)
         {
