+++ src/Raven.Server/Raven.Server.csproj
@@ -111,6 +111,7 @@
   <ItemGroup>
     <ProjectReference Include="..\Raven.Client\Raven.Client.csproj" />
     <ProjectReference Include="..\Sparrow.Server\Sparrow.Server.csproj" />
+    <ProjectReference Include="..\Sparrow\Sparrow.csproj" />
     <ProjectReference Include="..\Voron\Voron.csproj" />
   </ItemGroup>
   <ItemGroup>
